refactor: move generate_eval_samples to sampling module
The eval sampler is a sampling utility — its sibling generate_batch
already lives in tao_diffusion/sampling.py. Keeping them side by side
makes the design difference (full PPO trajectory vs portable eval x_0)
obvious to future readers.

The function is unchanged; only its location moved. train.py now imports
it alongside generate_batch. Updated the docstring to cross-reference
generate_batch.
